+++ app/storage/storage.js
@@ -1,29 +1,11 @@
-const { BlobServiceClient, generateBlobSASQueryParameters, BlobSASPermissions, StorageSharedKeyCredential } = require('@azure/storage-blob')
-const { connectionString, accountName, accountKey, applicationDocumentsStorage } = require('../config/storage')
-const containerName = applicationDocumentsStorage
-
-const getApplicationUrl = async (sbi, reference) => {
+const { applicationDocumentsStorage } = require('../config/storage')
+const getApplicationUrl = async (sbi, applicationReference) => {
   try {
-    const blobName = `${sbi}/${reference}.pdf`
-    const blobServiceClient = BlobServiceClient.fromConnectionString(connectionString)
-    const containerClient = blobServiceClient.getContainerClient(containerName)
-    const blobClient = containerClient.getBlobClient(blobName)
-
-    const sharedKeyCredential = new StorageSharedKeyCredential(accountName, accountKey)
-    const sasToken = generateBlobSASQueryParameters({
-      containerName,
-      blobName,
-      permissions: BlobSASPermissions.parse('r'),
-      startsOn: new Date(),
-      expiresOn: new Date(new Date().valueOf() + 3600 * 1000 * 2) // 2 hour (allows for time differences)
-    }, sharedKeyCredential).toString()
-
-    const applicationUrl = `${blobClient.url}?${sasToken}`
-
+    const blobName = `${sbi}/${applicationReference}.pdf`
+    const applicationUrl = `/download-application/${applicationDocumentsStorage}/${blobName}`
     return applicationUrl
   } catch (error) {
     console.error('Error: application URL', error)
-    // throw error
     return 'urlError'
   }
 }
