+++ test/unit/app/plugins/router.test.js
@@ -38,7 +38,7 @@ describe('routes plugin test', () => {
       routePaths.push(element.path)
     })
     expect(routePaths).toEqual([
-      '/check-details', '/cookies', '/healthy', '/healthz', '/signin-oidc', '/update-details', '/vet-visits', '/', '/assets/{path*}', '/check-details', '/cookies'
+      '/check-details', '/cookies', '/healthy', '/healthz', '/signin-oidc', '/update-details', '/vet-visits', '/', '/assets/{path*}', '/download-application/{containerName}/{blobName}', '/check-details', '/cookies'
     ])
   })
 })
